+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@
 
 * #10 Upgrade to Bootique 0.20
 * #11 Basic Derby-based unit test framework
+* #12 Port JDBC test utils from Cayenne
 
 ## 0.11